[settings] uncheck settings if not compatible with target
I wanted to make the line unsensitive, but it doesn't seems to be
possible. Either all the lines or nothing. As a fallback, the settings
is unselected. It is not perfect (some stuff might be missed, you can
check it back, ...) and the settings are removed at generation time
anyway.
Not sure it can be done better unless someone finds the way to control
the list element sensitivity one by one.